refactor(sana-wm/dit): split monolith into sana_wm.py + sana_wm_components.py
Mirror mickqian's two-file DiT layout. sana_wm_components.py now holds the
reusable pieces (RMSNorm/ShortConv primitives, WanRotaryPosEmbed + RoPE helpers,
UCPE camera geometry, chunk/attention utils, GDN scan kernels, embedders, FFN,
_downscale_to_reference_rms, the attention class, MultiHeadCrossAttention) plus
the cache-slot constants and the mutable _SANA_WM_TRITON_* fallback globals that
the attention class owns. sana_wm.py keeps SanaWMBlock + SanaWMTransformer3DModel
+ EntryClass and re-exports every component symbol so tests/stages that import
from this module path are unchanged.

Pure code relocation — no behavior change. CPU gate unchanged (113 pass / 2
pre-existing fail). Model class name kept (checkpoint config _class_name compat).
